refactor(Auth/usermanagers): put limitationshelper in a trait
