+++ includes/controllers/UserController.php
@@ -13,11 +13,14 @@ use YesWiki\Core\Service\DbService;
 use YesWiki\Core\Service\PageManager;
 use YesWiki\Core\Service\TripleStore;
 use YesWiki\Core\Service\UserManager;
+use YesWiki\Core\Trait\LimitationsTrait;
 use YesWiki\Core\YesWikiController;
 use YesWiki\Security\Controller\SecurityController;
 
 class UserController extends YesWikiController
 {
+    use LimitationsTrait;
+
     public const DEFAULT_NAME_MAX_LENGTH = 80;
     public const DEFAULT_EMAIL_MAX_LENGTH = 254;
     
@@ -389,26 +392,5 @@ class UserController extends YesWikiController
             throw new Exception(_t('USER_THIS_IS_NOT_A_VALID_EMAIL').".");
         }
         return $value;
-    }
-
-    /**
-     * init and store limitations in limitations array
-     * @param string $parameterName
-     * @param string $limitationKey
-     * @param mixed $type
-     * @param mixed $default
-     * @param string $errorMessageKey
-     */
-    private function initLimitationHelper(string $parameterName, string $limitationKey, $type, $default, string $errorMessageKey)
-    {
-        $this->limitations[$limitationKey] = $default;
-        if ($this->params->has($parameterName)) {
-            $parameter = $this->params->get($parameterName);
-            if (!filter_var($parameter, FILTER_VALIDATE_INT)) {
-                trigger_error(_t($errorMessageKey));
-            } else {
-                $this->limitations[$limitationKey] = $parameter;
-            }
-        }
     }
 }
